add flag & counting files based on dir/basename

+++ fileRenamer/main.go
@@ -1,55 +1,77 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"io/fs"
 	"log"
 	"os"
 	"path/filepath"
+	"sort"
 	"strconv"
 	"strings"
 )
 
-func main() {
-	var toRename []string
-	fileCount := 0
-	dir := "sample"
+type matchResult struct {
+	base  string
+	index int
+	ext   string
+}
 
-	err := filepath.Walk(dir, func(path string, info fs.FileInfo, err error) error {
+func main() {
+	toRename := make(map[string][]string)
+	var walkDir string
+	var dryrun bool
+	flag.StringVar(&walkDir, "walkdir", "sample", "the directory where you save all the files that need to rename")
+	flag.BoolVar(&dryrun, "dry", true, "whether or not should the tool change the names")
+	flag.Parse()
+
+	err := filepath.Walk(walkDir, func(path string, info fs.FileInfo, err error) error {
 		if err != nil {
 			return fmt.Errorf("failed access path: %s ", path)
 		}
 		if info.IsDir() {
 			return nil
 		}
-		if _, err := match(info.Name(), 0); err == nil {
-			fileCount++
-			toRename = append(toRename, info.Name())
+		if res, err := match(info.Name()); err == nil {
+			curDir := filepath.Dir(path)
+			key := filepath.Join(curDir, fmt.Sprintf("%s.%s", res.base, res.ext))
+			toRename[key] = append(toRename[key], info.Name())
 		}
 		return nil
 	})
 	if err != nil {
-		log.Fatal("Failed to walk through the dir", dir)
+		log.Fatal("Failed to walk through the dir", walkDir)
 	}
 
-	for _, originName := range toRename {
-		newName, err := match(originName, fileCount)
-		if err != nil {
-			log.Fatal("no match ", err)
-		}
-		newPath := filepath.Join(dir, newName)
-		oldPath := filepath.Join(dir, originName)
+	for key, files := range toRename {
+		dir := filepath.Dir(key)
+		count := len(files)
+		sort.Strings(files)
+		for i, filename := range files {
 
-		err = os.Rename(oldPath, newPath)
-		if err != nil {
-			log.Fatal("failed to rename ", err)
+			res, err := match(filename)
+			if err != nil {
+				log.Fatal("no match", err)
+			}
+
+			newName := fmt.Sprintf("%s-%d of %d.%s", res.base, (i + 1), count, res.ext)
+			newPath := filepath.Join(dir, newName)
+			oldPath := filepath.Join(dir, filename)
+
+			fmt.Printf("mv %s => %s\n", oldPath, newPath)
+			if !dryrun {
+				err = os.Rename(oldPath, newPath)
+				if err != nil {
+					log.Fatal("failed to rename", err)
+				}
+			}
 		}
-		fmt.Printf("mv %s => %s\n", oldPath, newPath)
 	}
 }
 
 //match returns the new file name, or an error if the file name did't match our pattern.
-func match(filename string, total int) (string, error) {
+func match(filename string) (*matchResult, error) {
 	pieces := strings.Split(filename, ".")
 	ext := pieces[len(pieces)-1]
 	tmp := strings.Join(pieces[0:len(pieces)-1], ".")
@@ -57,8 +79,7 @@ func match(filename string, total int) (string, error) {
 	name := strings.Join(pieces[0:len(pieces)-1], "_")
 	number, err := strconv.Atoi(pieces[len(pieces)-1])
 	if err != nil {
-		return "", fmt.Errorf("%s did not match our pattern", filename)
+		return nil, fmt.Errorf("%s did not match our pattern", filename)
 	}
-
-	return fmt.Sprintf("%s - %d of %d.%s", strings.Title(name), number, total, ext), nil
+	return &matchResult{strings.Title(name), number, ext}, nil
 }
